@@ -1,4 +1,4 @@
-// +build !ignore_autogenerated
+//go:build !ignore_autogenerated
 
 // Code generated by controller-gen. DO NOT EDIT.
 
